requirements.txt
asgiref==3.4.1
cffi==1.14.6
colorama==0.4.4
cryptography==3.4.8
Django==3.2.7
django-extensions==3.1.3
django-sslserver==0.22
Pillow==8.3.2
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
qrcode==7.3
six==1.16.0
sqlparse==0.4.1
Werkzeug==2.0.1